requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.4
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
